utils: Also move fixupUserHomeProperty() to Os
As this is the only place it is used now and feels like a good fit. Also
make it a regular function for simplicity, now that it is contained in a
class.
