+++ utils/src/main/kotlin/Utils.kt
@@ -22,7 +22,6 @@
 package org.ossreviewtoolkit.utils
 
 import java.io.File
-import java.lang.IllegalArgumentException
 import java.net.URI
 import java.net.URISyntaxException
 import java.security.Permission
@@ -76,29 +75,6 @@ val ortDataDirectory by lazy {
     }?.let {
         File(it)
     } ?: Os.userHomeDirectory.resolve(".ort")
-}
-
-/**
- * Check if the "user.home" property is set to a sane value and otherwise set it to the value of an (OS-specific)
- * environment variable for the user home directory, and return that value. This works around the issue that esp. in
- * certain Docker scenarios "user.home" is set to "?", see https://bugs.openjdk.java.net/browse/JDK-8193433 for some
- * background information.
- */
-inline fun <reified T : Any> T.fixupUserHomeProperty(): String {
-    val userHome = System.getProperty("user.home")
-    if (!userHome.isNullOrBlank() && userHome != "?") return userHome
-
-    val fallbackUserHome = listOfNotNull(
-        Os.env["HOME"],
-        Os.env["USERPROFILE"]
-    ).firstOrNull {
-        it.isNotBlank()
-    } ?: throw IllegalArgumentException("Unable to determine a user home directory.")
-
-    log.info { "Fixing up the user home directory from '$userHome' to '$fallbackUserHome'." }
-    System.setProperty("user.home", fallbackUserHome)
-
-    return fallbackUserHome
 }
 
 /**
